ensure missed playerstate events get played
Sometimes certain playerstate events would not get played if prediction
is on.
To reproduce:
1. Player needs to have a high ping (>= 150ms works)
2. Player stands on an item that is about to respawn (e.g. a weapon)
3. Player fires LG continously to produce lots of playerstate
   events (EV_FIRE_WEAPON)
4. Item respawns and is picked up immediately, but the pickup event
   (EV_ITEM_PICKUP) will not be triggered by cgame (and thus the item
   pickup sound will not play)

The item pickup cannot be predicted in this case because the client
doesn't know the item is about to respawn in advance. Because the event
buffer (which only holds two events) is filled quickly by other events,
the error does not get corrected.
Code to buffer more predicted events and check if any old events were
changed already existed, but was commented out (probably because it did
not work properly if cg_optimizePrediction was on).
This patch *should* make it work with cg_optimizePrediction on or off.